// src/lib/enums/partials.enum.ts
export enum INPUT_SIZE {
    BIG = 'big',
    SMALL = 'small'
}

export enum INPUT_POSITION {
    START = 'start',
    END = 'end'
}

export enum BUTTON_SIZE {
    BIG = 'big',
    SMALL = 'small'
}

export enum BUTTON_TYPE {
    PRIMARY = 'primary',
    SECONDARY = 'secondary',
    TERTIARY = 'tertiary',
}

export enum ARIA_LABEL {
    DISPLAY_PASSWORD = 'display the password',
    HIDE_PASSWORD = 'hide the password'
}

export enum AVATAR_SIZE {
    SMALL = "31px",
    MEDIUM = "42px",
    MEDIUM_LARGE = "68px",
    LARGE = "128px",
    EXTRA_LARGE = "180px"
}

export const AVATAR_ICON_SIZE = {
    SMALL: {
        width: "20px",
        height: "24px"
    },
    MEDIUM: {
        width: "26.76px",
        height: "31.29px"
    },
    MEDIUM_LARGE: {
        width: "38px",
        height: "38px"
    },
    LARGE: {
        width: "81.57px",
        height: "95.37px"
    },
    EXTRA_LARGE: {
        width: "94px",
        height: "93px"
    }
};